Funcionando Glide con imagenes provisionales. Comenzando cambio del RecyclerViewActivity to Fragment
