+++ app/src/main/java/com/example/bookpedia/Temporal_provider.kt
@@ -2,18 +2,24 @@ package com.example.bookpedia
 
 class Temporal_provider {
     companion object{
-        val Books_List = listOf<Book>(
+        val Books_List = listOf(
             Book(
                 "El Despertar de Leviatán",
                 "James SA Corey",
-                "02/06/11",
-                "https://www.google.com/search?q=Despertar+de+leviatan+fecha+de+publicaci%C3%B3n&client=opera-gx&sxsrf=ALiCzsaA0V0lSyDlV6RX16yIEivQlCKKHA:1652238238856&tbm=isch&source=iu&ictx=1&vet=1&fir=KRe2KJTIOvgwDM%252CUrwTiPUGdYLAsM%252C%252Fm%252F0k3nj9m&usg=AI4_-kStbj-sCJKrXcUbOYPkAgBVp31eBQ&sa=X&ved=2ahUKEwiLrb6Ku9b3AhX1omoFHY7iB30Q_B16BAgNEAE#imgrc=KRe2KJTIOvgwDM"
+                "02/06/2011",
+                "https://images-na.ssl-images-amazon.com/images/I/51CYoSkO+zL._SY344_BO1,204,203,200_.jpg"
             ),
             Book(
                 "Harry Potter y la Piedra Filosofal",
                 "J.K. Rowling",
-                "26/06/97",
-            "https://www.google.com/search?q=La+piedra+filosofal+Fecha+de+publicaci%C3%B3n+libro&client=opera-gx&sxsrf=ALiCzsYM3Bpm3CzQ-yFF6SnCVPezzl5RWg:1652238350640&tbm=isch&source=iu&ictx=1&vet=1&fir=S63LgcQYfYw4uM%252Cu9Q9-O9ggnfqOM%252C%252Fm%252F0c_vk&usg=AI4_-kRItzEHG86Jt3R_xQf98Evfl7FWDg&sa=X&ved=2ahUKEwjNlOW_u9b3AhW3rmoFHW3PCfMQ_B16BAgREAE#imgrc=S63LgcQYfYw4uM"
+                "26/06/1997",
+                "https://i.ibb.co/kxW0h4R/51cgf-Db-86-S-SY344-BO1-204-203-200-QL70-ML2.jpg"
+            ),
+            Book(
+                "La Guerra de Calibán",
+            "James SA Corey",
+                    "04/10/2017",
+                "https://i.ibb.co/dG9tyFm/content.jpg"
             )
         )
     }
