+++ app/src/main/res/layout/activity_main.xml
@@ -21,10 +21,4 @@
         app:headerLayout="@layout/nav_header_main"
         app:menu="@menu/activity_main_drawer"
         android:layout_gravity="start"/>
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recyclerView"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="start"
-    />
 </androidx.drawerlayout.widget.DrawerLayout>
